LRAC-12641 Move components used by AppliedFilters to filter/shared component
